Improved error handling! Also created commandParsers, so more code is out of the main segment! Made a second set of commands, adminCommands, for commands that should require some sort of permission
