+++ CTFNoteTaker.py
@@ -7,14 +7,12 @@ import re
 import socket
 import string
 import sqlite3
+import traceback
 from Utilities.StringUtils import *
 from Utilities.conf import *
-from commands.startnew import *
-from commands.listchals import *
-from commands.listctfs import *
-from commands.create import *
-from commands.add import *
-from commands.read import *
+from commandParser import *
+from adminCommandParser import *
+
 def handle():
     print("GETTING BUFFER\n")
     global readbuffer
@@ -32,45 +30,10 @@ def handle():
             cmd=line[3]
             cmd=cmd[1:]
             user=(line[0])[1:line[0].index("!")]
-            if(cmd==".help"):
-                if(len(line)>4):
-                    if(line[4]=="public"):
-                        help_chan()
-                    else:
-                        help(user)
-                else:
-                    help(user)
-
-            elif(cmd==".startnew"):
-                startnewcmd(line,user,s,conn,c)
-
-
-            elif(cmd==".listchal"):
-                list_chals(line,s,c)
-
-
-            elif(cmd==".listctf"):
-                list_CTFs(c,s)
-
-            elif(cmd==".create"):
-                startcreatecmd(s,line,conn,c)
-
-
-            elif(cmd==".add"):
-                startaddcmd(s,c,conn,user,line)
-
-
-            elif(cmd==".read"):
-                startreadcmd(user,s,c,line)
-
-            elif(cmd==".joinfail"):#pm to bot if it connects but doesnt join channel
-                join_chan_if_it_fails()
-
-            elif(cmd==".quit"):
-                if(len(line)>4):
-                    quit(line[4])
-                else:
-                    help_chan()
+            if(cmd.startswith(".")):
+                cmdParser(s,c,conn,user,line,cmd)
+            elif(cmd.startswith("~")):
+                adminCmdParser(s,c,conn,user,line,cmd)
 
 
 def is_first_run():
@@ -90,69 +53,9 @@ def is_first_run():
     else:
         return False
 
-
-
-
-
-def check_pass(password):
-    passfile=open("password","r").read()
-    passfile=passfile.replace("\n","")
-    if(password==passfile):
-        return True
-    else:
-        return False
-
 def pingpong(pong):
     s.send(bytes("PONG %s\r\n" % pong,"UTF-8"))
     print("PONG\n")
-
-def help_chan():
-    printChan(s,".help , .startnew <CTF> , .listchal <CTF> , .listctf , .create <CTF> <chalname> , .add <CTF> <chalname> <note>, .read <CTF> <chalname> , .quit <pass>")
-    s.send(bytes("PRIVMSG %s :Do not use whitespace in CTF name(you can use it in challenge name)\r\n" % CHAN,"UTF-8"))
-    s.send(bytes("PRIVMSG %s :Prefix your note with \"note:\" (without quotes)\r\n" % CHAN,"UTF-8"))
-    print("HELP\n")
-
-def help(user):
-    printUser(s,".help , .startnew <CTF> , .listchal <CTF> , .listctf , .create <CTF> <chalname> , .add <CTF> <chalname> <note>, .read <CTF> <chalname> , .quit <pass>",user)
-    printUser(s,"Do not use whitespace in CTF name(you can use it in challenge name)", user)
-    printUser(s,"Prefix your note with \"note:\" (without quotes)", user)
-    print("HELP\n")
-
-
-
-
-
-
-def create(CTF,challenge):
-    try:
-        challenge=" ".join(challenge)
-        c.execute("INSERT INTO challenges(title,ctfID) VALUES((?),(SELECT ctfID FROM ctf WHERE name=(?)))",(challenge,CTF))
-        conn.commit()
-        s.send(bytes("PRIVMSG %s :added challenge %s to %s\r\n" % (CHAN , challenge , CTF),"UTF-8"))
-    except:
-        s.send(bytes("PRIVMSG %s :CTF doesnt exist , if you are certain it does spam NETWORKsecurity\r\n" % CHAN,"UTF-8"))
-
-
-
-
-def join_chan_if_it_fails():
-    #patch because it doesnt join each time
-    s.send(bytes("JOIN %s\r\n" % CHAN,"UTF-8"))
-
-def quit(password):
-
-    if(check_pass(password)):
-        s.send(bytes("QUIT :\r\n","UTF-8"))
-
-        conn.close()
-        sys.exit()
-    else:
-        print("FAILED QUIT\n")
-
-
-
-
-
 
 sqlite_file = "database.sqlite"
 conn = sqlite3.connect(sqlite_file)
@@ -178,6 +81,9 @@ while(1):
     try:
         handle()
     except:
-        printChan(s,sys.exc_info()[0])
-        printUser(s,sys.exc_info()[0],MASTER)
-        print("BZZZZZZZZZZZZZZZTTTTT *******    ERROR   *** " + sys.exc_info()[0])
+        print("BZZZZZZZZZZZZZZZTTTTT *******    ERROR   *** " + str(sys.exc_info()[0]))
+        if("Keyboard" in str(sys.exc_info()[0])):
+            printChan(s, NICK + " is shutting down.")
+            break
+        printChan(s,"An internal error occured!")
+        printUser(s,"Error " + str(sys.exc_info()[0]),MASTER)
