+++ Utilities/conf.py
@@ -7,3 +7,4 @@ REALNAME = "Hal Glados"
 MASTER = "CTFNoteTaker" #NO HUMAN WILL RULE ME!
 
 BANNEDPHRASEMSG = "You used a banned phrase!"
+ADMINCOMMANDMODE = "MASTER"
