+++ adminCommands/error.py
@@ -1,0 +1,11 @@
+import sys
+import re
+import socket
+import string
+import sqlite3
+from Utilities.conf import *
+from Utilities.StringUtils import *
+from commands.create import *
+
+def throwErrorAdminCMD(s,c,conn,user,line):
+    raise Exception("I raised an error! " + str(line))
